Move barcode widget logic to separate package
