+++ datacapture/src/main/java/com/google/android/fhir/datacapture/views/barcode/mlkit/md/ScopedExecutor.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.mlkit.md
+package com.google.android.fhir.datacapture.views.barcode.mlkit.md
 
 import java.util.concurrent.Executor
 import java.util.concurrent.atomic.AtomicBoolean
