+++ datacapture/src/main/java/com/google/android/fhir/datacapture/views/barcode/mlkit/md/barcodedetection/BarcodeResultFragment.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.mlkit.md.barcodedetection
+package com.google.android.fhir.datacapture.views.barcode.mlkit.md.barcodedetection
 
 import android.content.DialogInterface
 import android.os.Bundle
@@ -28,8 +28,8 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.google.android.fhir.datacapture.R
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-import com.google.mlkit.md.camera.WorkflowModel
-import com.google.mlkit.md.camera.WorkflowModel.WorkflowState
+import com.google.android.fhir.datacapture.views.barcode.mlkit.md.camera.WorkflowModel
+import com.google.android.fhir.datacapture.views.barcode.mlkit.md.camera.WorkflowModel.WorkflowState
 
 /** Displays the bottom sheet to present barcode fields contained in the detected barcode. */
 class BarcodeResultFragment : BottomSheetDialogFragment() {
