+++ datacapture/src/main/java/com/google/android/fhir/datacapture/views/barcode/mlkit/md/camera/CameraSourcePreview.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.mlkit.md.camera
+package com.google.android.fhir.datacapture.views.barcode.mlkit.md.camera
 
 import android.content.Context
 import android.util.AttributeSet
@@ -24,7 +24,7 @@ import android.view.SurfaceView
 import android.widget.FrameLayout
 import com.google.android.fhir.datacapture.R
 import com.google.android.gms.common.images.Size
-import com.google.mlkit.md.Utils
+import com.google.android.fhir.datacapture.views.barcode.mlkit.md.Utils
 import java.io.IOException
 
 /** Preview the camera image in the screen. */
