+++ datacapture/src/main/java/com/google/android/fhir/datacapture/views/barcode/mlkit/md/camera/FrameMetadata.kt
@@ -14,16 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.mlkit.md.camera
+package com.google.android.fhir.datacapture.views.barcode.mlkit.md.camera
 
-import java.nio.ByteBuffer
-
-/** An interface to process the input camera frame and perform detection on it. */
-interface FrameProcessor {
-
-  /** Processes the input frame with the underlying detector. */
-  fun process(data: ByteBuffer, frameMetadata: FrameMetadata, graphicOverlay: GraphicOverlay)
-
-  /** Stops the underlying detector and release resources. */
-  fun stop()
-}
+/** Metadata info of a camera frame. */
+class FrameMetadata(val width: Int, val height: Int, val rotation: Int)
