+++ datacapture/src/main/java/com/google/android/fhir/datacapture/views/barcode/mlkit/md/camera/FrameProcessorBase.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.mlkit.md.camera
+package com.google.android.fhir.datacapture.views.barcode.mlkit.md.camera
 
 import android.os.SystemClock
 import android.util.Log
@@ -22,11 +22,9 @@ import androidx.annotation.GuardedBy
 import com.google.android.gms.tasks.OnFailureListener
 import com.google.android.gms.tasks.Task
 import com.google.android.gms.tasks.TaskExecutors
-import com.google.mlkit.md.CameraInputInfo
-import com.google.mlkit.md.InputInfo
-import com.google.mlkit.md.ScopedExecutor
-import com.google.mlkit.md.addOnFailureListener
-import com.google.mlkit.md.addOnSuccessListener
+import com.google.android.fhir.datacapture.views.barcode.mlkit.md.CameraInputInfo
+import com.google.android.fhir.datacapture.views.barcode.mlkit.md.InputInfo
+import com.google.android.fhir.datacapture.views.barcode.mlkit.md.ScopedExecutor
 import com.google.mlkit.vision.common.InputImage
 import java.nio.ByteBuffer
