+++ datacapture/src/main/res/layout/camera_preview_overlay.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
 >
 
-    <com.google.mlkit.md.camera.GraphicOverlay
+    <com.google.android.fhir.datacapture.views.barcode.mlkit.md.camera.GraphicOverlay
         android:id="@+id/camera_preview_graphic_overlay"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
